+++ src/Validators/StringValidator.php
@@ -1,15 +1,24 @@
 <?php
+declare(strict_types=1);
+
 namespace Fynix\Validators;
 
 use Fynix\ValidationError;
 
 class StringValidator extends LengthValidatorBase {
-    public function __construct(
+    protected function __construct(
         string $name,
         string $propertyName)
     { 
         parent::__construct($name, $propertyName);
-        $this->length(2, 50);
+        $this->minLength = 2;
+        $this->maxLength = 50;
+    }
+
+    /** @internal */
+    public static function __makeInternal(string $name, string $propertyName): static
+    {
+        return new static($name, $propertyName);
     }
 
     public function validate(mixed $fieldValue) : ?ValidationError